+++ packages/flyer_chat_file_message/lib/src/flyer_chat_file_message.dart
@@ -153,7 +153,6 @@ class FlyerChatFileMessage extends StatelessWidget {
         return Row(
           mainAxisSize: MainAxisSize.min,
           crossAxisAlignment: CrossAxisAlignment.end,
-          textBaseline: TextBaseline.alphabetic,
           children: [fileContent, const SizedBox(width: 4), timeAndStatus],
         );
       case TimeAndStatusPosition.end: