+++ packages/flyer_chat_text_stream_message/lib/src/text_segment.dart
@@ -1,0 +1,20 @@
+import 'package:flutter/animation.dart';
+
+sealed class TextSegment {
+  final String text;
+  const TextSegment(this.text);
+}
+
+class StaticSegment extends TextSegment {
+  StaticSegment(super.text);
+}
+
+class AnimatingSegment extends TextSegment {
+  final AnimationController controller;
+  final Animation<double> fadeAnimation;
+  AnimatingSegment(super.text, this.controller, this.fadeAnimation);
+
+  void dispose() {
+    controller.dispose();
+  }
+}